+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Fixed
+
+### Changed
+## [0.0.19] - 2025-07-18
+
+### Added
+
 - Added duplicate weather check cleanup
 - Added next reset date to GDD model card on dashboard
 - Added bespoke weed pressure calculation and chart